+++ src/test/java/com/kurkus/kusinsa/service/user/DeviceServiceTest.java
@@ -1,0 +1,4 @@
+import static org.junit.jupiter.api.Assertions.*;
+class DeviceServiceTest {
+  
+}
